# Improve the performance of `ConfOctoTransport#setSources`