While doing large conference testing, we noticed that this method was a big hotspot because of the way we were redundantly iterating over endpoints and their sources.  This PR reworks the method a bit to build everything in a single pass.  I wrote a quick test to compare the old code vs the new code:

These numbers are based on 10 iterations of calls to `ConfOctoTransport#setSources` with 500 eps, each with one audio and one video source:

Old code:
```
Took 1193 ms
Took 749 ms
Took 638 ms
Took 512 ms
Took 621 ms
Took 415 ms
Took 344 ms
Took 530 ms
Took 342 ms
Took 415 ms
```

New code:
```
Took 558 ms
Took 258 ms
Took 215 ms
Took 169 ms
Took 164 ms
Took 139 ms
Took 160 ms
Took 200 ms
Took 142 ms
Took 148 ms
```

I left a couple intermediary commits in to show a couple things, but they'll be squashed when merging.